remove mentions function
